Added homework for lecture 9
