+++ src/main/java/lecture9/homework/service/songs/SongService.java
@@ -1,0 +1,18 @@
+package homework.service.songs;
+
+import homework.repository.song.Song;
+
+import java.util.List;
+
+public interface SongService {
+
+  void addSong(Song song);
+
+  void deleteSong(int id);
+
+  Song getSong(int id);
+
+  List<Song> getAllSongs();
+
+  void update(Song song);
+}
