+++ src/main/java/lecture9/homework/service/users/UserService.java
@@ -1,0 +1,20 @@
+package homework.service.users;
+
+import homework.repository.user.User;
+
+import java.util.List;
+
+public interface UserService {
+
+    void addUser(User user);
+
+    void deleteUser(int id);
+
+    User getUser(int id);
+
+    List<User> getAllUsers();
+
+    void update(User user);
+
+    void playSong(int id, int songId);
+}
