--- src/main/lecture9.homework/jbdc/repository/artist/Artist.java
@@ -1,17 +1,0 @@
-package jbdc.repository.artist;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class Artist {
-
-    private int id;
-    private String firstName;
-    private String lastName;
-    private String genre;
-    private LocalDate dateOfBirth;
-
-
-}
