--- src/main/lecture9.homework/jbdc/repository/user/User.java
@@ -1,16 +1,0 @@
-package jbdc.repository.user;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class User {
-
-    private int id;
-    private String firstName;
-    private String lastName;
-    private LocalDate dateOfBirth;
-    private boolean isPaid;
-
-}
