--- src/main/lecture9.homework/jbdc/service/SongServiceImpl.java
@@ -1,49 +1,0 @@
-package jbdc.service;
-
-
-import jbdc.repository.song.Song;
-import jbdc.repository.song.SongRepository;
-import jbdc.service.songs.SongService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-
-public class SongServiceImpl implements SongService {
-
-    public final SongRepository songRepository;
-
-    public SongServiceImpl(final SongRepository songRepository) {
-        this.songRepository = songRepository;
-    }
-
-    @Override
-    public void addSong(Song song) {
-        songRepository.add(song);
-    }
-
-    @Override
-    public void deleteSong(int id) {
-        songRepository.delete(id);
-
-    }
-
-    @Override
-    public Song getSong(int id) {
-        return songRepository.get(id);
-    }
-
-    @Override
-    public List<Song> getAllSongs() {
-        return songRepository.getAll();
-    }
-
-    @Override
-    public void update(Song song) {
-        songRepository.update(song);
-    }
-
-}
